Ignoring empty fields during mapping
Addresses https://github.com/hxl-team/HXLator/issues/61

Also, if a property that has already been mapped is mapped again, the
old value will be replaced. This allows users to fix mistakes they may
have made, but it is somewhat limiting in that we cannot produce any
datasets that have more than one value per property. We might need to
address this once we have cardinalities in the standard.